Example Repository schema and row definitions in one file
